+++ src/util/kebab2camel.js
@@ -10,7 +10,7 @@
  * @return {string}
  */
 function kebab2camel(source) {
-    return source.replace(/-([a-z])/g, function (match, alpha) {
+    return source.replace(/-([a-z])/ig, function (match, alpha) {
         return alpha.toUpperCase();
     });
 }